__tests__/transfer.test.ts
import { getUserInput } from "../transfer";

describe("Transfer Function Tests", () => {
    describe("Incorrect Behavior", () => {
        it("should return incorrect results", async () => {
            jest.spyOn(process.stdin, "emit").mockImplementationOnce(value => {
                process.stdin.emit("data", "0.01\n"); 
                return true;
            });

            const userInput = await getUserInput();
            expect(userInput).toBeCloseTo(0.01); 

            jest.restoreAllMocks();
        });
    });

    describe("Security Issues", () => {
        it("should only allow access to permitted accounts", () => {
            const senderPublicKey = "senderPublicKey";
            const recipientPublicKey = "recipientPublicKey";
            expect(senderPublicKey).toEqual("senderPublicKey");
            expect(recipientPublicKey).toEqual("recipientPublicKey");
        });
    });

    const validateAmount = (amount: any): boolean => typeof amount === 'number' && amount >= 0.01 && amount <= 1;

    describe("Invalid Arguments", () => {
        it("should handle invalid input arguments", () => {
            const isValid = (input: any) => validateAmount(input);
            expect(isValid(0.5)).toBe(true);
            expect(isValid('invalid')).toBe(false);
        });
    });

    describe("Behavior Prevention", () => {
        it("should prevent transactions from underfunded accounts", async () => {
            expect(0).toBeLessThan(5000); 
        });
    });

    describe("Error Handling", () => {
        it("should handle exceptions properly", async () => {
            const throwException = async () => { throw new Error("Simulated error"); };
            await expect(throwException()).rejects.toThrow("Simulated error");
        });
    });
});
